# Parallelize traversing AST's in the language server