--- a/sway-lsp/Cargo.toml
+++ b/sway-lsp/Cargo.toml
@@ -19,6 +19,7 @@ notify-debouncer-mini = { version = "0.2.0" }
 parking_lot = "0.12.1"
 proc-macro2 = "1.0.5"
 quote = "1.0.9"
+rayon = "1.5.0"
 ropey = "1.2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.60"