--- a/sway-lsp/src/traverse/lexed_tree.rs
+++ b/sway-lsp/src/traverse/lexed_tree.rs
@@ -2,6 +2,7 @@ use crate::{
     core::token::{AstToken, SymbolKind, Token},
     traverse::{Parse, ParseContext},
 };
+use rayon::iter::{IntoParallelRefIterator, ParallelBridge, ParallelIterator};
 use sway_ast::{
     expr::LoopControlFlow, ty::TyTupleDescriptor, Assignable, CodeBlockContents, ConfigurableField,
     Expr, ExprArrayDescriptor, ExprStructField, ExprTupleDescriptor, FnArg, FnArgs, FnSignature,
@@ -15,16 +16,16 @@ use sway_types::{Ident, Span, Spanned};
 
 pub fn parse(lexed_program: &LexedProgram, ctx: &ParseContext) {
     insert_module_kind(ctx, &lexed_program.root.tree.kind);
-    for item in &lexed_program.root.tree.items {
-        item.value.parse(ctx);
-    }
+    lexed_program.root.tree.items.par_iter().for_each(|item| item.value.parse(ctx));
 
-    for (.., dep) in &lexed_program.root.submodules {
-        insert_module_kind(ctx, &dep.module.tree.kind);
-        for item in &dep.module.tree.items {
-            item.value.parse(ctx);
-        }
-    }
+    lexed_program
+        .root
+        .submodules
+        .par_iter()
+        .for_each(|(_, dep)| {
+            insert_module_kind(ctx, &dep.module.tree.kind);
+            dep.module.tree.items.par_iter().for_each(|item| item.value.parse(ctx));
+        });
 }
 
 fn insert_module_kind(ctx: &ParseContext, kind: &ModuleKind) {
@@ -102,9 +103,7 @@ impl Parse for Expr {
                 args.get().address.parse(ctx);
             }
             Expr::Struct { fields, .. } => {
-                for expr in fields.get() {
-                    expr.parse(ctx);
-                }
+                fields.get().into_iter().par_bridge().for_each(|field| field.parse(ctx));
             }
             Expr::Tuple(tuple) => {
                 tuple.get().parse(ctx);
@@ -137,10 +136,10 @@ impl Parse for Expr {
             } => {
                 insert_keyword(ctx, match_token.span());
                 value.parse(ctx);
-                for branch in branches.get() {
+                branches.get().into_iter().par_bridge().for_each(|branch| {
                     branch.pattern.parse(ctx);
                     branch.kind.parse(ctx);
-                }
+                });
             }
             Expr::While {
                 while_token,
@@ -153,9 +152,7 @@ impl Parse for Expr {
             }
             Expr::FuncApp { func, args } => {
                 func.parse(ctx);
-                for expr in args.get().into_iter() {
-                    expr.parse(ctx);
-                }
+                args.get().into_iter().par_bridge().for_each(|expr| expr.parse(ctx));
             }
             Expr::Index { target, arg } => {
                 target.parse(ctx);
@@ -169,13 +166,9 @@ impl Parse for Expr {
             } => {
                 target.parse(ctx);
                 if let Some(contract_args) = contract_args_opt {
-                    for expr in contract_args.get().into_iter() {
-                        expr.parse(ctx);
-                    }
+                    contract_args.get().into_iter().par_bridge().for_each(|expr| expr.parse(ctx));
                 }
-                for expr in args.get().into_iter() {
-                    expr.parse(ctx);
-                }
+                args.get().into_iter().par_bridge().for_each(|expr| expr.parse(ctx));
             }
             Expr::FieldProjection { target, .. } => {
                 target.parse(ctx);
@@ -257,6 +250,7 @@ impl Parse for ItemStruct {
         self.fields
             .get()
             .into_iter()
+            .par_bridge()
             .for_each(|field| field.value.parse(ctx));
     }
 }
@@ -275,6 +269,7 @@ impl Parse for ItemEnum {
         self.fields
             .get()
             .into_iter()
+            .par_bridge()
             .for_each(|field| field.value.parse(ctx));
     }
 }
@@ -299,7 +294,7 @@ impl Parse for ItemTrait {
 
         self.trait_items
             .get()
-            .iter()
+            .par_iter()
             .for_each(|annotated| match &annotated.value {
                 sway_ast::ItemTraitItem::Fn(fn_sig, _) => fn_sig.parse(ctx),
                 sway_ast::ItemTraitItem::Const(item_const, _) => item_const.parse(ctx),
@@ -309,7 +304,7 @@ impl Parse for ItemTrait {
         if let Some(trait_defs_opt) = &self.trait_defs_opt {
             trait_defs_opt
                 .get()
-                .iter()
+                .par_iter()
                 .for_each(|item| item.value.parse(ctx));
         }
     }
@@ -331,7 +326,7 @@ impl Parse for ItemImpl {
 
         self.contents
             .get()
-            .iter()
+            .par_iter()
             .for_each(|item| match &item.value {
                 ItemImplItem::Fn(fn_decl) => fn_decl.parse(ctx),
                 ItemImplItem::Const(const_decl) => const_decl.parse(ctx),
@@ -345,7 +340,7 @@ impl Parse for ItemAbi {
 
         self.abi_items
             .get()
-            .iter()
+            .par_iter()
             .for_each(|annotated| match &annotated.value {
                 sway_ast::ItemTraitItem::Fn(fn_sig, _) => fn_sig.parse(ctx),
                 sway_ast::ItemTraitItem::Const(item_const, _) => item_const.parse(ctx),
@@ -355,7 +350,7 @@ impl Parse for ItemAbi {
         if let Some(abi_defs_opt) = self.abi_defs_opt.as_ref() {
             abi_defs_opt
                 .get()
-                .iter()
+                .par_iter()
                 .for_each(|item| item.value.parse(ctx));
         }
     }
@@ -385,6 +380,7 @@ impl Parse for ItemStorage {
         self.fields
             .get()
             .into_iter()
+            .par_bridge()
             .for_each(|field| field.value.parse(ctx));
     }
 }
@@ -403,6 +399,7 @@ impl Parse for ItemConfigurable {
         self.fields
             .get()
             .into_iter()
+            .par_bridge()
             .for_each(|field| field.value.parse(ctx));
     }
 }
@@ -429,9 +426,9 @@ impl Parse for UseTree {
     fn parse(&self, ctx: &ParseContext) {
         match self {
             UseTree::Group { imports } => {
-                for use_tree in imports.get().into_iter() {
+                imports.get().into_iter().par_bridge().for_each(|use_tree| {
                     use_tree.parse(ctx);
-                }
+                });
             }
             UseTree::Rename { as_token, .. } => {
                 insert_keyword(ctx, as_token.span());
@@ -491,7 +488,7 @@ impl Parse for FnArgs {
     fn parse(&self, ctx: &ParseContext) {
         match self {
             FnArgs::Static(punct) => {
-                punct.into_iter().for_each(|fn_arg| fn_arg.parse(ctx));
+                punct.into_iter().par_bridge().for_each(|fn_arg| fn_arg.parse(ctx));
             }
             FnArgs::NonStatic {
                 self_token,
@@ -507,7 +504,7 @@ impl Parse for FnArgs {
                     insert_keyword(ctx, mut_token.span());
                 }
                 if let Some((.., punct)) = args_opt {
-                    punct.into_iter().for_each(|fn_arg| fn_arg.parse(ctx));
+                    punct.into_iter().par_bridge().for_each(|fn_arg| fn_arg.parse(ctx));
                 }
             }
         }
@@ -523,9 +520,7 @@ impl Parse for FnArg {
 
 impl Parse for CodeBlockContents {
     fn parse(&self, ctx: &ParseContext) {
-        for statement in self.statements.iter() {
-            statement.parse(ctx);
-        }
+        self.statements.par_iter().for_each(|statement| statement.parse(ctx));
         if let Some(expr) = self.final_expr_opt.as_ref() {
             expr.parse(ctx);
         }
@@ -564,9 +559,7 @@ impl Parse for ExprArrayDescriptor {
     fn parse(&self, ctx: &ParseContext) {
         match self {
             ExprArrayDescriptor::Sequence(punct) => {
-                for expr in punct.into_iter() {
-                    expr.parse(ctx);
-                }
+                punct.into_iter().par_bridge().for_each(|expr| expr.parse(ctx));
             }
             ExprArrayDescriptor::Repeat { value, length, .. } => {
                 value.parse(ctx);
@@ -629,14 +622,10 @@ impl Parse for Pattern {
                 }
             }
             Pattern::Constructor { args, .. } | Pattern::Tuple(args) => {
-                for pattern in args.get().into_iter() {
-                    pattern.parse(ctx);
-                }
+                args.get().into_iter().par_bridge().for_each(|pattern| pattern.parse(ctx));
             }
             Pattern::Struct { fields, .. } => {
-                for field in fields.get().into_iter() {
-                    field.parse(ctx);
-                }
+                fields.get().into_iter().par_bridge().for_each(|field| field.parse(ctx));
             }
             _ => {}
         }
@@ -680,9 +669,7 @@ impl Parse for ExprTupleDescriptor {
     fn parse(&self, ctx: &ParseContext) {
         if let ExprTupleDescriptor::Cons { head, tail, .. } = self {
             head.parse(ctx);
-            for expr in tail.into_iter() {
-                expr.parse(ctx);
-            }
+            tail.into_iter().par_bridge().for_each(|expr| expr.parse(ctx));
         }
     }
 }
@@ -691,9 +678,7 @@ impl Parse for TyTupleDescriptor {
     fn parse(&self, ctx: &ParseContext) {
         if let TyTupleDescriptor::Cons { head, tail, .. } = self {
             head.parse(ctx);
-            for expr in tail.into_iter() {
-                expr.parse(ctx);
-            }
+            tail.into_iter().par_bridge().for_each(|expr| expr.parse(ctx));
         }
     }
 }