## Description
This PR introduces using rayon's parallel iterators when traversing the lexed and parsed AST's. I initially extended this to also occur on the typed AST but was running into deadlock issues with how DashMap works. 

closes #4959

| Request Type/Method    | Previous  | New      | Improvement % |
|------------------------|-----------|----------|---------------|
| compile_and_traverse   | 2.7732s   | 2.5632s  | 7.57%         |

**edit**

I've split the compilation and traversal stages so they can be benchmarked independently. Here are the actual traversal performance results.

| Request Type/Method    | Previous  | New      | Improvement % |
|------------------------|-----------|----------|---------------|
| traverse   | 422ms   | 211ms  | 50.0%         |

## Checklist

- [x] I have linked to any relevant issues.
- [ ] I have commented my code, particularly in hard-to-understand areas.
- [ ] I have updated the documentation where relevant (API docs, the reference, and the Sway book).
- [ ] I have added tests that prove my fix is effective or that my feature works.
- [ ] I have added (or requested a maintainer to add) the necessary `Breaking*` or `New Feature` labels where relevant.
- [ ] I have done my best to ensure that my PR adheres to [the Fuel Labs Code Review Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [ ] I have requested a review from the relevant team or maintainers.