@xunilrj 2023-08-28
A 5% gain is giving me the impression we are parallelizing too much...
Maybe just per module will yield better results? I think there are too much `par_iter()`...