@JoshuaBatty 2023-08-29
> Maybe just per module will yield better results? I think there are too much `par_iter()`

Thanks for the tip, I have been trying with a variety of situations. Only traversing in parallel per module didn't lead to any difference in performance. I'll keep investigating.