@JoshuaBatty 2023-08-31
Ok this should be ready now. Pre allocating the DashMap gave us a few extra % performance win as well.