@JoshuaBatty 2023-08-31
I think this benchmark is a bit misleading as the compiler performance is bundled with the traversal perf. I'm going to put up a PR that splits compilation and traversal into unique stages so we measure each stage separately.